+++ packages/ui/src/components/TimelineItem/TimelineItem.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { screen } from '@testing-library/react';
 
-import { renderWithTheme } from '../../utils/renderWithTheme';
+import { renderWithTheme } from '../../testHelpers';
 import { TimelineItem } from './TimelineItem';
 
 describe('TimelineItem', () => {
